Finished moving all the definitions
