+++ app/definitions.py
@@ -4,9 +4,15 @@ SMS_TYPE = 'sms'
 EMAIL_TYPE = 'email'
 LETTER_TYPE = 'letter'
 
+TEMPLATE_TYPES = [SMS_TYPE, EMAIL_TYPE, LETTER_TYPE]
+
+NOTIFICATION_TYPE = [EMAIL_TYPE, SMS_TYPE, LETTER_TYPE]
+
 # NOTIFICATION PRIORITIES
 NORMAL = 'normal'
 PRIORITY = 'priority'
+
+TEMPLATE_PROCESS_TYPE = [NORMAL, PRIORITY]
 
 
 # BRANDING DEFINITIONS
@@ -25,9 +31,20 @@ INTERNATIONAL_SMS_TYPE = 'international_sms'
 INBOUND_SMS_TYPE = 'inbound_sms'
 SCHEDULE_NOTIFICATIONS = 'schedule_notifications'
 
+SERVICE_PERMISSION_TYPES = [
+    EMAIL_TYPE,
+    SMS_TYPE,
+    LETTER_TYPE,
+    INTERNATIONAL_SMS_TYPE,
+    INBOUND_SMS_TYPE,
+    SCHEDULE_NOTIFICATIONS
+]
+
 # WHITELIST TYPES
 MOBILE_TYPE = 'mobile'
 EMAIL_TYPE = 'email'
+
+WHITELIST_RECIPIENT_TYPE = [MOBILE_TYPE, EMAIL_TYPE]
 
 
 # API KEY TYPEs
@@ -39,6 +56,10 @@ KEY_TYPE_TEST = 'test'
 MMG_PROVIDER = "mmg"
 FIRETEXT_PROVIDER = "firetext"
 SES_PROVIDER = 'ses'
+
+SMS_PROVIDERS = [MMG_PROVIDER, FIRETEXT_PROVIDER]
+EMAIL_PROVIDERS = [SES_PROVIDER]
+PROVIDERS = SMS_PROVIDERS + EMAIL_PROVIDERS
 
 
 # JOB STATUSES
@@ -63,6 +84,50 @@ NOTIFICATION_TECHNICAL_FAILURE = 'technical-failure'
 NOTIFICATION_TEMPORARY_FAILURE = 'temporary-failure'
 NOTIFICATION_PERMANENT_FAILURE = 'permanent-failure'
 
+NOTIFICATION_STATUS_TYPES_FAILED = [
+    NOTIFICATION_TECHNICAL_FAILURE,
+    NOTIFICATION_TEMPORARY_FAILURE,
+    NOTIFICATION_PERMANENT_FAILURE,
+]
+
+NOTIFICATION_STATUS_TYPES_COMPLETED = [
+    NOTIFICATION_SENT,
+    NOTIFICATION_DELIVERED,
+    NOTIFICATION_FAILED,
+    NOTIFICATION_TECHNICAL_FAILURE,
+    NOTIFICATION_TEMPORARY_FAILURE,
+    NOTIFICATION_PERMANENT_FAILURE,
+]
+
+NOTIFICATION_STATUS_SUCCESS = [
+    NOTIFICATION_SENT,
+    NOTIFICATION_DELIVERED
+]
+
+NOTIFICATION_STATUS_TYPES_BILLABLE = [
+    NOTIFICATION_SENDING,
+    NOTIFICATION_SENT,
+    NOTIFICATION_DELIVERED,
+    NOTIFICATION_FAILED,
+    NOTIFICATION_TECHNICAL_FAILURE,
+    NOTIFICATION_TEMPORARY_FAILURE,
+    NOTIFICATION_PERMANENT_FAILURE,
+]
+
+NOTIFICATION_STATUS_TYPES = [
+    NOTIFICATION_CREATED,
+    NOTIFICATION_SENDING,
+    NOTIFICATION_SENT,
+    NOTIFICATION_DELIVERED,
+    NOTIFICATION_PENDING,
+    NOTIFICATION_FAILED,
+    NOTIFICATION_TECHNICAL_FAILURE,
+    NOTIFICATION_TEMPORARY_FAILURE,
+    NOTIFICATION_PERMANENT_FAILURE,
+]
+
+NOTIFICATION_STATUS_TYPES_NON_BILLABLE = list(set(NOTIFICATION_STATUS_TYPES) - set(NOTIFICATION_STATUS_TYPES_BILLABLE))
+
 
 # USES STATUSES
 INVITED_USER_STATUS_TYPES = ['pending', 'accepted', 'cancelled']
@@ -77,3 +142,16 @@ SEND_LETTERS = 'send_letters'
 MANAGE_API_KEYS = 'manage_api_keys'
 PLATFORM_ADMIN = 'platform_admin'
 VIEW_ACTIVITY = 'view_activity'
+
+# List of permissions
+PERMISSION_LIST = [
+    MANAGE_USERS,
+    MANAGE_TEMPLATES,
+    MANAGE_SETTINGS,
+    SEND_TEXTS,
+    SEND_EMAILS,
+    SEND_LETTERS,
+    MANAGE_API_KEYS,
+    PLATFORM_ADMIN,
+    VIEW_ACTIVITY
+]
