+++ app/models.py
@@ -32,23 +32,7 @@ from app import (
 from app.history_meta import Versioned
 from app.utils import convert_utc_time_in_bst, convert_bst_to_utc
 
-from app.definitions import (
-    SMS_TYPE,
-    EMAIL_TYPE,
-    LETTER_TYPE,
-    NORMAL,
-    PRIORITY,
-    INTERNATIONAL_SMS_TYPE, INBOUND_SMS_TYPE, SCHEDULE_NOTIFICATIONS, DVLA_ORG_HM_GOVERNMENT, BRANDING_GOVUK,
-    MOBILE_TYPE, FIRETEXT_PROVIDER, MMG_PROVIDER, SES_PROVIDER, JOB_STATUS_FINISHED, JOB_STATUS_IN_PROGRESS,
-    JOB_STATUS_PENDING, JOB_STATUS_SCHEDULED, JOB_STATUS_CANCELLED, JOB_STATUS_ERROR,
-    JOB_STATUS_SENDING_LIMITS_EXCEEDED, JOB_STATUS_SENT_TO_DVLA, JOB_STATUS_READY_TO_SEND, NOTIFICATION_DELIVERED,
-    NOTIFICATION_TECHNICAL_FAILURE, NOTIFICATION_TEMPORARY_FAILURE, NOTIFICATION_PERMANENT_FAILURE, NOTIFICATION_FAILED,
-    NOTIFICATION_SENT, NOTIFICATION_SENDING, NOTIFICATION_PENDING, NOTIFICATION_CREATED, INVITED_USER_STATUS_TYPES,
-    MANAGE_TEMPLATES, MANAGE_USERS, MANAGE_SETTINGS, SEND_LETTERS, VIEW_ACTIVITY, SEND_TEXTS, PLATFORM_ADMIN,
-    MANAGE_API_KEYS, SEND_EMAILS)
-
-TEMPLATE_TYPES = [SMS_TYPE, EMAIL_TYPE, LETTER_TYPE]
-TEMPLATE_PROCESS_TYPE = [NORMAL, PRIORITY]
+from app.definitions import *
 
 template_types = db.Enum(*TEMPLATE_TYPES, name='template_type')
 
@@ -142,16 +126,6 @@ class DVLAOrganisation(db.Model):
     __tablename__ = 'dvla_organisation'
     id = db.Column(db.String, primary_key=True)
     name = db.Column(db.String(255), nullable=True)
-
-
-SERVICE_PERMISSION_TYPES = [
-    EMAIL_TYPE,
-    SMS_TYPE,
-    LETTER_TYPE,
-    INTERNATIONAL_SMS_TYPE,
-    INBOUND_SMS_TYPE,
-    SCHEDULE_NOTIFICATIONS
-]
 
 
 class ServicePermissionTypes(db.Model):
@@ -256,8 +230,8 @@ class ServicePermission(db.Model):
         return '<{} has service permission: {}>'.format(self.service_id, self.permission)
 
 
-WHITELIST_RECIPIENT_TYPE = [MOBILE_TYPE, EMAIL_TYPE]
 whitelist_recipient_types = db.Enum(*WHITELIST_RECIPIENT_TYPE, name='recipient_type')
+
 
 class ServiceWhitelist(db.Model):
     __tablename__ = 'service_whitelist'
@@ -450,11 +424,6 @@ class TemplateHistory(db.Model):
         return serialized
 
 
-SMS_PROVIDERS = [MMG_PROVIDER, FIRETEXT_PROVIDER]
-EMAIL_PROVIDERS = [SES_PROVIDER]
-PROVIDERS = SMS_PROVIDERS + EMAIL_PROVIDERS
-
-NOTIFICATION_TYPE = [EMAIL_TYPE, SMS_TYPE, LETTER_TYPE]
 notification_types = db.Enum(*NOTIFICATION_TYPE, name='notification_type')
 
 
@@ -614,50 +583,6 @@ class VerifyCode(db.Model):
     def check_code(self, cde):
         return check_hash(cde, self._code)
 
-
-NOTIFICATION_STATUS_TYPES_FAILED = [
-    NOTIFICATION_TECHNICAL_FAILURE,
-    NOTIFICATION_TEMPORARY_FAILURE,
-    NOTIFICATION_PERMANENT_FAILURE,
-]
-
-NOTIFICATION_STATUS_TYPES_COMPLETED = [
-    NOTIFICATION_SENT,
-    NOTIFICATION_DELIVERED,
-    NOTIFICATION_FAILED,
-    NOTIFICATION_TECHNICAL_FAILURE,
-    NOTIFICATION_TEMPORARY_FAILURE,
-    NOTIFICATION_PERMANENT_FAILURE,
-]
-
-NOTIFICATION_STATUS_SUCCESS = [
-    NOTIFICATION_SENT,
-    NOTIFICATION_DELIVERED
-]
-
-NOTIFICATION_STATUS_TYPES_BILLABLE = [
-    NOTIFICATION_SENDING,
-    NOTIFICATION_SENT,
-    NOTIFICATION_DELIVERED,
-    NOTIFICATION_FAILED,
-    NOTIFICATION_TECHNICAL_FAILURE,
-    NOTIFICATION_TEMPORARY_FAILURE,
-    NOTIFICATION_PERMANENT_FAILURE,
-]
-
-NOTIFICATION_STATUS_TYPES = [
-    NOTIFICATION_CREATED,
-    NOTIFICATION_SENDING,
-    NOTIFICATION_SENT,
-    NOTIFICATION_DELIVERED,
-    NOTIFICATION_PENDING,
-    NOTIFICATION_FAILED,
-    NOTIFICATION_TECHNICAL_FAILURE,
-    NOTIFICATION_TEMPORARY_FAILURE,
-    NOTIFICATION_PERMANENT_FAILURE,
-]
-
-NOTIFICATION_STATUS_TYPES_NON_BILLABLE = list(set(NOTIFICATION_STATUS_TYPES) - set(NOTIFICATION_STATUS_TYPES_BILLABLE))
 
 NOTIFICATION_STATUS_TYPES_ENUM = db.Enum(*NOTIFICATION_STATUS_TYPES, name='notify_status_type')
 
@@ -982,20 +907,6 @@ class InvitedUser(db.Model):
     # play nice with marshmallow yet
     def get_permissions(self):
         return self.permissions.split(',')
-
-
-# List of permissions
-PERMISSION_LIST = [
-    MANAGE_USERS,
-    MANAGE_TEMPLATES,
-    MANAGE_SETTINGS,
-    SEND_TEXTS,
-    SEND_EMAILS,
-    SEND_LETTERS,
-    MANAGE_API_KEYS,
-    PLATFORM_ADMIN,
-    VIEW_ACTIVITY
-]
 
 
 class Permission(db.Model):
