+++ app/service/utils.py
@@ -1,12 +1,14 @@
 import itertools
 
 from app.models import (
-    ServiceWhitelist,
-    MOBILE_TYPE, EMAIL_TYPE
+    ServiceWhitelist
 )
 from app.definitions import (
     KEY_TYPE_TEST,
-    KEY_TYPE_NORMAL
+    KEY_TYPE_NORMAL,
+    KEY_TYPE_TEAM,
+    MOBILE_TYPE,
+    EMAIL_TYPE
 )
 
 from notifications_utils.recipients import allowed_to_send_to
